requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
httpx